// src/api/models/ImageInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tgs_v2
 * Timeguessr Social API V2
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ImageInfo
 */
export interface ImageInfo {
  /**
   *
   * @type {number}
   * @memberof ImageInfo
   */
  latitude: number;
  /**
   *
   * @type {number}
   * @memberof ImageInfo
   */
  longitude: number;
  /**
   *
   * @type {string}
   * @memberof ImageInfo
   */
  country: string;
  /**
   *
   * @type {number}
   * @memberof ImageInfo
   */
  year: number;
  /**
   *
   * @type {string}
   * @memberof ImageInfo
   */
  imageDescription: string;
  /**
   *
   * @type {string}
   * @memberof ImageInfo
   */
  imageUrl: string;
}

/**
 * Check if a given object implements the ImageInfo interface.
 */
export function instanceOfImageInfo(value: object): value is ImageInfo {
  if (!("latitude" in value) || value["latitude"] === undefined) return false;
  if (!("longitude" in value) || value["longitude"] === undefined) return false;
  if (!("country" in value) || value["country"] === undefined) return false;
  if (!("year" in value) || value["year"] === undefined) return false;
  if (!("imageDescription" in value) || value["imageDescription"] === undefined)
    return false;
  if (!("imageUrl" in value) || value["imageUrl"] === undefined) return false;
  return true;
}

export function ImageInfoFromJSON(json: any): ImageInfo {
  return ImageInfoFromJSONTyped(json, false);
}

export function ImageInfoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ImageInfo {
  if (json == null) {
    return json;
  }
  return {
    latitude: json["latitude"],
    longitude: json["longitude"],
    country: json["country"],
    year: json["year"],
    imageDescription: json["imageDescription"],
    imageUrl: json["imageUrl"],
  };
}

export function ImageInfoToJSON(json: any): ImageInfo {
  return ImageInfoToJSONTyped(json, false);
}

export function ImageInfoToJSONTyped(
  value?: ImageInfo | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    latitude: value["latitude"],
    longitude: value["longitude"],
    country: value["country"],
    year: value["year"],
    imageDescription: value["imageDescription"],
    imageUrl: value["imageUrl"],
  };
}