// src/api/models/RoundResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tgs_v2
 * Timeguessr Social API V2
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface RoundResult
 */
export interface RoundResult {
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  roundIndex: number;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  totalPoints: number;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  locationPoints: number;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  timePoints: number;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  yearsOff: number;
  /**
   *
   * @type {string}
   * @memberof RoundResult
   */
  distanceOff: string;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  latitude: number;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  longitude: number;
  /**
   *
   * @type {number}
   * @memberof RoundResult
   */
  guessedYear: number;
}

/**
 * Check if a given object implements the RoundResult interface.
 */
export function instanceOfRoundResult(value: object): value is RoundResult {
  if (!("roundIndex" in value) || value["roundIndex"] === undefined)
    return false;
  if (!("totalPoints" in value) || value["totalPoints"] === undefined)
    return false;
  if (!("locationPoints" in value) || value["locationPoints"] === undefined)
    return false;
  if (!("timePoints" in value) || value["timePoints"] === undefined)
    return false;
  if (!("yearsOff" in value) || value["yearsOff"] === undefined) return false;
  if (!("distanceOff" in value) || value["distanceOff"] === undefined)
    return false;
  if (!("latitude" in value) || value["latitude"] === undefined) return false;
  if (!("longitude" in value) || value["longitude"] === undefined) return false;
  if (!("guessedYear" in value) || value["guessedYear"] === undefined)
    return false;
  return true;
}

export function RoundResultFromJSON(json: any): RoundResult {
  return RoundResultFromJSONTyped(json, false);
}

export function RoundResultFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): RoundResult {
  if (json == null) {
    return json;
  }
  return {
    roundIndex: json["roundIndex"],
    totalPoints: json["totalPoints"],
    locationPoints: json["locationPoints"],
    timePoints: json["timePoints"],
    yearsOff: json["yearsOff"],
    distanceOff: json["distanceOff"],
    latitude: json["latitude"],
    longitude: json["longitude"],
    guessedYear: json["guessedYear"],
  };
}

export function RoundResultToJSON(json: any): RoundResult {
  return RoundResultToJSONTyped(json, false);
}

export function RoundResultToJSONTyped(
  value?: RoundResult | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    roundIndex: value["roundIndex"],
    totalPoints: value["totalPoints"],
    locationPoints: value["locationPoints"],
    timePoints: value["timePoints"],
    yearsOff: value["yearsOff"],
    distanceOff: value["distanceOff"],
    latitude: value["latitude"],
    longitude: value["longitude"],
    guessedYear: value["guessedYear"],
  };
}